LayoutTests/accessibility/gtk/aria-busy-changed-notification-expected.txt

This tests that changing the aria-busy value results in a state-changed notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axElement.boolAttributeValue('AXElementBusy') is false
Setting aria-busy to true on listbox.
Setting aria-busy to false on listbox.
AXElementBusyChanged: AXTitle:  AXRole: AXWebArea
AXElementBusyChanged: AXTitle:  AXRole: AXListBox
AXElementBusyChanged: AXTitle:  AXRole: AXListBox
PASS axElement.boolAttributeValue('AXElementBusy') is false
PASS successfullyParsed is true

TEST COMPLETE
